refactor: Implementa viewModel do componente Select
